+++ packages/utils-media/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dovenv/utils-media
 
+## 1.5.28
+
+### Patch Changes
+
+- Add support in predocs for build only parts
+
+- Updated dependencies []:
+  - @dovenv/utils@1.5.28
+
 ## 1.5.27
 
 ### Patch Changes